Added support for Google Analytics
