+++ edelphi/src/main/webapp/WEB-INF/web.xml
@@ -24,11 +24,11 @@
     <url-pattern>/*</url-pattern>
   </filter-mapping>
   <filter>
-    <filter-name>DefaultsFilter</filter-name>
-    <filter-class>fi.metatavu.edelphi.DefaultsFilter</filter-class>
+    <filter-name>CommonsFilter</filter-name>
+    <filter-class>fi.metatavu.edelphi.CommonsFilter</filter-class>
   </filter>
   <filter-mapping>
-    <filter-name>DefaultsFilter</filter-name>
+    <filter-name>CommonsFilter</filter-name>
     <url-pattern>/*</url-pattern>
 </filter-mapping>
   <filter>
