+++ edelphi/src/main/webapp/jsp/templates/generic_head.jsp
@@ -141,3 +141,5 @@
     }
   }
 </script>
+
+<jsp:include page="/jsp/templates/google_analytics.jsp"></jsp:include>
